fix bug when loading people. changed Create user class into a module. reorganization of decorators.
