+++ Project/create_user.rb
@@ -1,8 +1,8 @@
 require_relative 'student'
 require_relative 'teacher'
 
-class CreateUser
-  def create_student(people)
+module CreateUser
+  def self.create_student(people)
     print 'Age: '
     age = gets.chomp.to_i
     print 'Name: '
@@ -16,7 +16,7 @@ class CreateUser
               end
   end
 
-  def create_teacher(people)
+  def self.create_teacher(people)
     print 'Age: '
     age = gets.chomp.to_i
     print 'Name: '
