+++ Project/data/people_data.rb
@@ -1,5 +1,7 @@
 require 'json'
+
 module PersonData
+
   def self.save_people(people)
     File.open('database/people.json', 'w') do |file|
       file.puts people.map { |person|
@@ -9,12 +11,17 @@ module PersonData
       }.to_json
     end
   end
+
   def self.load_people(people)
     return unless File.exist?('database/people.json')
     file = File.read('database/people.json')
     people_hash = JSON.parse(file)
     people_hash.each do |person|
-      people << Person.new(person['name'], person['id'])
+      if person['class'] == 'Student'
+        people << Student.new(person['id'], person['name'])
+      else
+        people << Teacher.new(person['id'], person['specialization'], name: person['name'])
+      end
     end
   end
 end
