+++ Project/database/people.json
@@ -1,1 +1,1 @@
-[{"name":763,"id":567,"class":"Person"},{"name":223,"id":637,"class":"Person"}]
+[{"name":"jimmy","id":726,"class":"Student"},{"name":"Unknown","id":802,"class":"Teacher"},{"name":"jimmy","id":38,"class":"Student"},{"name":"Carl","id":117,"class":"Teacher"}]
