+++ Project/person.rb
@@ -1,5 +1,5 @@
 require_relative 'nameable'
-require_relative 'capitalize_decorator'
+require_relative 'decorators/capitalize_decorator'
 require_relative 'rental'
 
 class Person < Nameable
